+++ src/bot/dcaHelper.ts
@@ -14158,7 +14158,9 @@ function createDCABotHelper<
           (settings.useSl &&
             (!settings.trailingSl || settings.useMultiSl) &&
             (settings.dealCloseConditionSL === CloseConditionEnum.tp ||
-              (settings.moveSL && d.deal.moveSlActivated))) ||
+              (settings.moveSL &&
+                d.deal.moveSlActivated &&
+                +(settings.slPerc ?? 0) === +(settings.moveSLValue ?? 0)))) ||
           (this.slAr &&
             settings.useSl &&
             settings.dealCloseConditionSL === CloseConditionEnum.dynamicAr)) &&
@@ -14207,7 +14209,9 @@ function createDCABotHelper<
         useSl &&
         (!trailingSl || useMultiSl) &&
         (dealCloseConditionSL === CloseConditionEnum.tp ||
-          (moveSL && d.deal.moveSlActivated))
+          (moveSL &&
+            d.deal.moveSlActivated &&
+            +(slPerc ?? 0) === +(settings.moveSLValue ?? 0)))
       ) {
         if (
           useMultiSl &&
@@ -15612,6 +15616,8 @@ function createDCABotHelper<
           useFixedSLPrices,
           fixedSlPrice,
           multiTp,
+          slPerc,
+          moveSLValue,
         } = await this.getAggregatedSettings(d.deal)
         const dealId = d.deal._id
         let closeBySl = true
@@ -15646,7 +15652,9 @@ function createDCABotHelper<
           useSl &&
           (!trailingSl || useMultiSl) &&
           (dealCloseConditionSL === CloseConditionEnum.tp ||
-            (moveSL && d.deal.moveSlActivated))
+            (moveSL &&
+              d.deal.moveSlActivated &&
+              +(slPerc ?? 0) === +(moveSLValue ?? 0)))
         ) {
           if (
             useMultiSl &&