+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [1.15.7] - 2026-04-20
+
+### Fixed
+
+- Move SL value got overwritten
+
+## [1.15.6] - 2026-04-13
+
+### Fixed
+
+- Long Wick types
+
+## [1.15.5] - 2026-04-13
+
+### Added
+
+- Timer to release mutex
+
 ## [1.15.4] - 2026-04-08
 
 ### Fixed